Move octree feature from shared project to Wpf.SharpDX
