+++ Source/HelixToolkit.SharpDX.Shared/Model/Interfaces.cs
@@ -1,0 +1,15 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="Interfaces.cs" company="Helix Toolkit">
+//   Copyright (c) 2014 Helix Toolkit contributors
+// </copyright>
+// --------------------------------------------------------------------------------------------------------------------
+
+using System;
+
+namespace HelixToolkit.Wpf.SharpDX
+{
+    public interface IGUID
+    {
+        Guid GUID { get; }
+    }
+}
